+++ web/package.json
@@ -38,7 +38,7 @@
     "typescript": "^5.3.3",
     "vite": "^4.5.2",
     "vite-plugin-tailwind-purgecss": "0.1.3",
-    "vite-plugin-wasm": "^3.3.0"
+    "vite-plugin-wasm-pack": "^0.1.12"
   },
   "type": "module",
   "dependencies": {
