--- web/src/lib/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "smithereens",
-  "version": "0.1.0",
-  "files": [
-    "smithereens_bg.wasm",
-    "smithereens.js",
-    "smithereens_bg.js",
-    "smithereens.d.ts"
-  ],
-  "module": "smithereens.js",
-  "types": "smithereens.d.ts",
-  "sideEffects": [
-    "./smithereens.js",
-    "./snippets/*"
-  ],
-  "dependencies": {
-    "vite-plugin-wasm": "^3.3.0"
-  }
-}
