--- web/src/lib/smithereens.d.ts
@@ -1,20 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-/**
-* @param {Peptidoglycan} precursor
-* @returns {string}
-*/
-export function pg_to_fragments(precursor: Peptidoglycan): string;
-/**
-*/
-export class Peptidoglycan {
-  free(): void;
-/**
-* @param {string} structure
-*/
-  constructor(structure: string);
-/**
-* @returns {string}
-*/
-  monoisotopic_mass(): string;
-}
