--- web/src/lib/smithereens.js
@@ -1,4 +1,0 @@
-import * as wasm from "./smithereens_bg.wasm";
-import { __wbg_set_wasm } from "./smithereens_bg.js";
-__wbg_set_wasm(wasm);
-export * from "./smithereens_bg.js";
