--- web/src/lib/smithereens_bg.wasm.d.ts
@@ -1,12 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-export const memory: WebAssembly.Memory;
-export function __wbg_peptidoglycan_free(a: number): void;
-export function peptidoglycan_new(a: number, b: number, c: number): void;
-export function peptidoglycan_monoisotopic_mass(a: number, b: number): void;
-export function pg_to_fragments(a: number, b: number): void;
-export function __wbindgen_add_to_stack_pointer(a: number): number;
-export function __wbindgen_malloc(a: number, b: number): number;
-export function __wbindgen_realloc(a: number, b: number, c: number, d: number): number;
-export function __wbindgen_free(a: number, b: number, c: number): void;
-export function __wbindgen_exn_store(a: number): void;
