+++ web/vite.config.ts
@@ -1,9 +1,10 @@
 import { purgeCss } from 'vite-plugin-tailwind-purgecss';
 import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
+import wasmPack from 'vite-plugin-wasm-pack';
 
 export default defineConfig({
-	plugins: [sveltekit(), purgeCss()],
+	plugins: [sveltekit(), purgeCss(), wasmPack('./smithereens')],
 	worker: {
 		format: 'es'
 	}
